reports views: Make code clearer, and consistencify
To begin with, especially the multiple_service_states was rendered in an
extremely innovative way, where the summary was printed in different
places, depending on the position of the sun, which rendered the summary
in slightly different ways, and where the logical beginning of the loop
was halfway through it, with many conditionals guarding the first and
last repetition into and out of that logical start.

The point of all of this was to make sure we render services in
inconsistent ways, depending on if we're creating a multi-service report
or a multi-servicegroup report. I don't see the point of trying to be
inconsistent - we're quite good at that without trying - so drop that
difference.

This also made me realize the difference between how hosts and services
are shown - hosts have summaries printed in the regular table, while
services have an extra table. When trying to make them look the same, I
decided the extra table is prettier, but I'm open to suggestions.

Wrap the host tables with report_block, so they're properly separated.

Also remove the nonsense-logic to exclude any included object with no
included state information - versions of this logic have been here for a
long time, but if this causes anything weird to happen, it's the input
that is broken. This fixes bug #6435.

(yes, consistencify is totally a word)

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>