report-options: Doesn't map 1:1 on Saved_reports_Model::get_report_info()
We want to rely on its information, but not all of it is within the
report options' scope, for example "id" and "username". Those keys may
be accessed by other parts of the code, querying for all of the data
(including the metadata) from a saved report (which means I won't change
Saved_reports_Model directly) but *we're* only interesting in the data
needed for input to reproduce a report.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>

Conflicts:
	application/libraries/Report_options.php

Yeah, nothing works without this, we need it in maint too.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>