reports/options: Validate {host,service}_filter_status
Since the values was an array, it passed the validation, but there's a
need for individual validation. This was pretty nasty since you could
pass arbitrarily data and store it in the DB, and if you're upgrading to
a < 6.0 version, that value gets unserialized.

In practice, this means that you could submit HTML containing a
serialized object, whose destructor (when unserialized in the upgrade
script) could contain too trustworthy code and become abused.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>