Kohana form helper: don't dubble escape
Putting your things in a safe makes your things unaccessable to outsiders, so they are safely stored. Putting that safe in another safe, makes the storage even safer.

Programming doesn't work that way, even if the developers of Kohana thinks it does. It just makes a totally mess... for the user... When creating an element, the form::attributes method calls the html::attributes method, which in turn escapes the data... Why do it before? The developer doesn't trust himself between the modules?

Futhermore... HTML has a feature. Each HTML tag can have named attributes. The attributes are named so you don't have to use indexes of the attributes, which is much harder to remembers and read than a name. Therefore the order of attributes doesn't matter.

Kohana-developers has appearently thought that sorting attributes was a good idéa, so form-attributes is in the beginning. So you have faster access to them? Doing this does two things:
 1. The server has to work harder rendering forms. Beacuse the server is fast we can waste it's resources to nessecary things like this... *sigh*
 2. The web browser stores the attributes in some kind of map for easy access. One possibly quite common storage backend for maps of strings is a binary tree. A balanced tree has the speed performance of O(log(n)). Adding elements sorted to a standard search tree makes the tree totally unbalanced (of not using a self-balancing tree...), the complexity of access is O(n)... But todays client computers is fast, lets use it!

Keep up the good work, Kohana!

Signed-off-by: Max Sikström <msikstrom@op5.com>