Report_options: Don't fail brutally as soon as there's an unknown option
To spot the error: open reports, select the report you want, and then -
just for shits-n-giggles - click the Select button. Watch your report
*not* be selected - or at least not your booleans.

The problem is a moronic white-list of cases where we're allowed to load
booleans, which is waaay to restricted.

This isn't so much a fix, as it's a workaround: all reports have more
than three configurables, so let's allow a submit button and some random
report type or something in addition to the report_id, and the bug won't
show up. Until it does, at which point I'll travel back in time and
smack myself - but since I haven't been smacked yet, I can only conclude
that this is fool-proof.

What we really need is to do something less stupid with our checkboxes,
but that'll have to wait.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>