Themes: removed (non-)support for themes
Themes is nice. You can design a tool with a theme of anything, lika rollercoste
r ride. You can visualize it using a flash with 3d. You can make a theme like a
MMORPG, where you walk around and collect borken services like acknowledgments,
and if you do it good, you can get "Achievement unlocked", and get to the next h
ostgroup. Or maybe write it like a novel, which makes you solve a tricky quest t
o se what the problem actually is. Have fun Sherlock Holmes!

Unfourtunatly we didn't support that. Or maybe it was what we did, when the them
e concept was to copy and replace half of Ninja. But most of our users whant to
do other whings with themes. For example configure the colors, logo, and add som
e branding. And make it easy, so our concept of themes was... let's say... broke
n by design.

So time just save our tab buttons so we don't need to tab through them<TAB>defau
<TAB> anymore, and do it... not maybe entirly right, but at least better, which
is, simply don't support making a copy of half ninja and try to understand the c
ode, and let them use skins instead, which is what they want to do.

Next up: fix skins to be able to do a little bit more...

Signed-off-by: Max Sikström <msikstrom@op5.com>