spec: Add php-process to build deps too
We run PHP code that requires this (such as PHP tap, and perhaps also
the code that was the reason we needed this in the regular dep list in
the first place) during build. So guess what: we depend on it.

Also bump the php tap version to one that tells you when it breaks due
to this. Stupid '@' characters *grumble*

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>