views/sla: Fix report group member listing
So, my mood just got a lot worse.

Because we generate what amounts to $n SLA reports in one when we select
$n {host,service}groups to include in the report (yet only 1 when we
select $n {hosts,services}), we can't use get_report_members(), because
it will return *all* hosts for *all* embedded SLA reports, but we're
iterating over the list of embedded SLA reports, so the result should be
limited to this one.

This one fixes that, but I'm sure it breaks some other case, and I'll
have to add yet another balancing act of if-else tango with a bunch of
is_array().

I need vodka.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>