Move code form Auth.php to User.php
Move some currently rarely-used authorization code from auth (which is
here supposed to mean "authentication") to user. While you could
make the argument that a user's authorization belong in
authorization.php, this is apparently not how this is designed.
Anywho... Organizing things this way makes mocking up rights much
easier, so go with this.

(This commit is imported from op5lib)

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>