commands: Make authorization checks a helluvalot more correct
When porting ninja to op5auth, we forgot to implement hostgroup rights,
servicegroup rights, and by-contact rights, even though making it
configurable in nacoma and saying it applied here kindof suggested
otherwise. Now all of the above exists, we handle "fun" commands that
mentions all kinds of types, and there's loads of tests to check the
logic. Don't you *dare* break this again and fail to discover it,
myself!

We remove the "if authed for host then authed for service" logic that
used to be here - livestatus will do that for us, if the user wants us
to, and won't if the user doesn't. Neat, huh?

This fixes bug #6546.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>