ListView: pluggable table renderers from modules
This is the last part to be able to, only by dropping a module to the module
directory, inject tables into the listview, to be able to present them and
search within them, like searching in all other tables.

This is a nessecary patch to, for example, convert trapper to a listview.

Signed-off-by: Max Sikström <msikstrom@op5.com>