Nagios_Auth: removed unused methods.
By some reason we had a couple of methods starting with "throw new Exception"
still there. I'm pretty sure noone of those were used, and if they were, a
deprectated-exception isn't much more of help than a, similar looking,
method-not-found-error for the end user. And for the developer, those say
approximatly the same thing.

Signed-off-by: Max Sikström <msikstrom@op5.com>