docs: Generate docs for modules
It could be useful especially for learning the syntax of the new
lsfilter-queries.

I got a hell of a lot of warnings when triggering 'make docs' locally
but the documentation came out just fine,
Documentation/html/classlistview.html even got some example code in
them, which is a good thing.

(And the @license tag is phpdoc, not doxygen)

Signed-off-by: Carl Helmertz <chelmertz@op5.com>