report_options: Remove report_id when creating "anonymous" reports
Formatting "anonymous" reports means "include all configuration, but
throw away any identifying information as we'll be replacing it later" -
we use it in particular for creating links to individual objects in
multi-object-reports. Anonymous reports did not get rid of the
report_id, which meant the report we did generate became a weird
mish-mash of the two reports, causing it to sometimes bail out with
a weird, unexpected "100% Undetermined" state. It also likely meant that
saving the subreport replaced the parent report, which is also quite
unexpected.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>