LSFilterList: Title text goes on link, not image
Ok, so it ended up becoming *and* image, but you do want it for the
link, so you can hover all over the link and get an explanation of what
the link points to. If you want an alternative text for an image, alt is
the attribute to use, not title.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>