Dojo js: Avoid annoying js error messages.
Whenever rules were saved in the synergy gui the query_for_state function put
out errors about not being able to split a string that was null.
Why, oh why was it considered a good idea to do a jquery selection on a title
attribute that renews every ten seconds?
Selection is now made with id. At least that is always found.

Signed-off-by: Henrik Smederod <hsmederod@op5.com>