auth/log: Notice-level logging now reveals *why* auth went bad
The auth lib needs to be a bit redesigned with error handling in mind,
it's tough to resolve what login() === false means: bad user, password,
bad driver logic etc; tough to resolve from the client code that is
(such as in the REST API, which now only responds 'You have to login'
when you feed it with bad credentials).

This patch makes it a bit easier to detect *what* went wrong.

(This commit is imported from op5lib)

Signed-off-by: Carl Helmertz <chelmertz@op5.com>