lsfilter: Make sorting really visible
Reusing the invented-by-someone-handsome 'lightblue' css color + visible
icons to indicate sorted-on and can-be-sorted-on columns. Triggering
sorting by clicking on all of the th which is now also hints about that
action by changing cursor and bgcolor on hover.

I favor visible as a design principle, instead of
everything-looks-so-slim-so-it-all-fits-together-and-nothing-sticks-out,
if you didn't notice.

This fixes #6894.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>