Unhandled Problems widget: Replace misleading '/' with helpful '+'
In the past, the unhandled problems widget showed the number of - for
example - the number of services in the unknown state in total, and then
additionally the number of services in the unknown state on non-OK
hosts.

As of recently, we instead show the number of services in the unknown
state on OK hosts, as well as the number of services in the unknown
state on non-OK hosts.

The new way probably makes more sense, as it very clearly shows you the
number of problems you're likely to care about. However, "4 Unhandled
problems / 98 on problem hosts" reads like "You've got 4 unhandled
problems in total, of which 98 are on problem hosts" which sounds
stupid. Writing "4 Unhandled problems + 98 on problem hosts" correctly
identifies the numbers as two different terms that make up the total.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>