ListView visualization: group statements close to top node better
a filter like [x] a and (b or c) and d and e should be grouped like:

(a) and (b or c) and (d and e)

so as few top nodes exists as possible, but keep the order, and keep and-nodes
at the top level. (a) becames a and filter group.

Signed-off-by: Max Sikström <msikstrom@op5.com>