csv: Add more tests, fix problems
One problem is that we always failed to print the report type (as in
report object type) for sla reports.

Another is that we always printed the host name twice - once on the
HOST_NAME column and once in the SERVICE_DESCRIPTION column before a
semi-colon - for avail reports. In other words, bug #6948

This resolves both those problems.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>