reports/common.js: Sort out handling of custom report periods
The problem was that it had about a dozen different, interesting cases,
and it almost, but not quite, created a unix timestamp, which is what
the backend wants, but then at the last second it changes its mind and
goes back to some other format. Sigh.

Also, it didn't work.

Remove most cases to only get one case left. This resolves bug #6958.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>