ListView: Fixed a couple of issues with floating header width
Also cleaned up a lot of ajax request code, and removed a couple of dynamically
created anonymous functions.

Anonymous functions is great in lot of cases, but when creating a anonymous
function, the surronding variable space is saved with the function, so those
variables is avalible when the function is runned. This can create memory-leak
issues. Removing those functions makes it lot easier to debug the code.

Signed-off-by: Max Sikström <msikstrom@op5.com>