ListView: make the width of the header correct in IE too
By some reason, a stray pixel was needed to get the width working i chrome, but
not IE.

By some stranger reason, the previous refactoring made that pixel nessecary in
IE too.

That is that pixel? why is is it needed? But when it's needed in both IE,
chrome and firefox, it should probably be there...

Signed-off-by: Max Sikström <msikstrom@op5.com>