Report_options: Create property initialization method
We tend to need for the sub-option-classes to modify the default
properties as generated by Report_options, but Report_options cannot
have it's defaults completely in the class definition, because PHP's
gettext doesn't support m_(). We used to do it after the constructor,
but if the constructor is called with an options argument, then
validation of those options might fail.

Create a new method for initializing properties, so you can inject
properties in a sub-class after the base class is setting up properties,
but before it starts to validate against them.

This makes the forever timeperiod work for summary reports, which the
included cuke test verifies.

There's a tiny CSS hack included that forces the graph image to be at
least a few pixels for that test to work.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>