filters: remove input elements from header original after cloning
The underlying header is (as far as I understand it) only there to
provide the proper dimensions for the clone. Thus, we can remove the
input field from the original in order to make the input fields
clickable from our cuke tests. It also eliminates duplicate id's for
those fields - amnd that's a good thing.

Signed-off-by: Anton Lofgren <alofgren@op5.com>