trend graphs: Stop enforcing height to 30 pixels
Enforcing it means that really long names (like, 30 characters
or so) will get wrapped and overwrite themselves, which isn't
very nice.

With this patch, people can use names as long as they like and
things will just wrap around neatly, with readable text.

If someone uses names where each individual word is longer than
what fits in the width of the text-field, it's still possible
to get a slightly crummy report, as parts of the text may be
invisible behind the bar. While that's sort of bad, it should
be a lot rarer than the entire text being too long and the
symptoms of it are by no means so severe that we need to cater
for it here and now.

Signed-off-by: Andreas Ericsson <ae@op5.se>