sla: Increase discoverability for "generate avail report from this sla"
Generating an avail report for this exact SLA report is a nice feature.
To understand that it existed, I had to read the code to notice that the
graph was linked and where it pointed.

By having the title text saying something more informative than "Uptime"
(yes, I know the bars display uptime, why is this hyperlinked? oh, it's
linked to another uptime view. right.) this will hopefully be slightly
less hard to understand.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>