Nagios_auth_Model: is no more...
After op5Auth was implemented, and ninja was ported to livestaus, the
Nagios_auth_Model was just a crazy designed wrapper for authorization, designed
as a signleton.

What could be done by just asking for the information was now done by stashing
the authorized user, dragging the permissions out of the user, stash those,
never lower the permissions in an instance, which coudln't be replaced, and
then ask livestatus once per object.

So now, the code is gone, and a big piece of history is removed, and well... it
looks like it works better, and also, making it possible to test things, like
limited users in reports.

Goodbye Nagios_auth_Model

Signed-off-by: Max Sikström <msikstrom@op5.com>