ORM generator: fetch key when asking for no keys in livestatus
WHen asking for no columns (which is an empty column list, not to be confused
with not asking for columns, which would fallback on all columns), the key columns
wasn't automatically added. The ORM interface is designed to having get_key() method
working for all objects, so this could be a problem when just expecting a list of
keys.

(This commit is imported from op5lib)

Signed-off-by: Max Sikström <msikstrom@op5.com>