reports/js: Only implement report deletion once
Bonus: only have one post-deletion behavior.

Before, avail and sla reports showed a growl box saying the report was
deleted, but kept the old report config was still there, and you had to
press "New" to get rid of it. The summary report showed a growl box for
about 0.0 seconds, before reloading the page with a clean report.

With this, we refresh with a new report for all report types, and we
don't even bother with trying to show a growl box that's not supposed to
be seen.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>