controllers/reports: SLA reports wrap Avail reports, not SLA
We treated the report as a avail report, except we used the SLA type. As
we need avail reports to support everything (relevant) that SLA can do
anyway in order to be able to click the SLA report to get an avail
report in the UI, we can just use an actual nested avail report.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>