tests: Increase sleep time when running recurring_downtime tests
A sleep time of one seconds apparently wasn't enough for the buildbot
environment when it was moody. Bump it, and stuff fail less. Fail bad.

Well, I'm sorry, but what do you expect me to do? I can't exactly wait
for the nagios command pipe to tell me that it accepted my command,
because pipes don't work that way.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>