reports model: Fix wildcard filters as per customer spec
Customers have reported that they like to search for host_name and/or
service_description as well as the general output. Since searching
free-style with wildcards is the only way to filter in particular
services when showing service status from a hostgroup list, we actually
need to put this back in.

While at it, we make the filters case insensitive (yes, in an oracle-
compatible way) and also make them handle "normal" fnmatch (glob)
style syntax.

This fixes issue #6342.

Signed-off-by: Max Sikström <msikstrom@@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>