nagstat helper process_macros(): Fall back to returning the string
In case we don't get enough data, it's always better to return the
unprocessed string than it is to return 'false'.

This undoes most of the damage introduced by the API-breaking
commit e16775688b3813f18e5bf4c9c782dc81f74804ea, which introduced a
new required variable to the function without corresponding updates
to all the widgets using the helpers and thereby partially fixes a
whole slew of bugs related to missing links and data.

In particular, it provides a partial fix for #6764.

Signed-off-by: Andreas Ericsson <ae@op5.se>