nagstat helper process_macros(): Use heuristics to determine object type
This function was broken for all callers when a new required variable
was introduced in commit e16775688b3813f18e5bf4c9c782dc81f74804ea,
but since both external and op5-supported widgets rely on the old
call-signature (two arguments required), we need to provide some
sort of fallback for the case where we don't get the newly introduced
third argument.

This patch provides that fallback, providing a backend fix for #6764.

Signed-off-by: Andreas Ericsson <ae@op5.se>