schedule: Remove formatting of report names in dropdown
It was all very half-hearted, where we only did it to avail reports
loaded on page load, and not for anything ajax-loaded. Apparently,
nobody missed it anywhere else.

I'm not sure what's the use of knowing if something already is scheduled
in the dropdown - if you want the info, it's more easily available down
in the table right below. Therefore, delete.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>