ListView: Add support for list compherensions and implode in custom columns
To get the comments from custom columns earlier, the only field actually having
the comments text was comments_with_info. This column contains id, author and
text, so it was impossible to just get a list of comments

This patch makes it possible to extract just text from the comments with:

[x[2] for x in comments_with_info]

Also, to convert an array to string, the implode function is useful.

implode(" | ", [x[2] for x in comments_with_info])

Signed-off-by: Max Sikström <msikstrom@op5.com>