Stylesheets: Well... actually care about variables defined in themes
There are a couple of ways to write sass files with templates and default
values.

Write a config file with default values for variables, write a common-css file
which handles detailed stylesheets, depending on the varibles. Include the
config in the beginning of the config file, so you make sure those variables
are set.

When writing the stylehset, you include either the common file in the
beginning. But then the all stylesheets are generated with the variables set.

Next way is to set the varibles before including the common template, so those
variables are set. But then, the common sass file includes config and resets
the variables.

From time to time, we managed both those ways... and quite often include the
default config several times, to make sure we really overwrite the theme
modifications.

The third way is to actually include the default variables, modify those that
should be modified, and then generate the css...

So for all of you missed the pink theme, it's back!

Signed-off-by: Max Sikström <msikstrom@op5.com>