reports/js: Handle case with empty success message
I have no idea what triggers it, but it sometimes happen for an hour or
two when running cuke tests, and then nothing works on that machine
until something cheers it up again, and meanwhile you'll have to use
another machine for testing.

My best guess is that we abort the request at an inopportune moment due
to sad synchronisation reasons.

Either way, handle a malformed response as an error, because it is.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>