report_test: Make tests take longer because architecture
When we load a config, we can't ask about timeperiod data until nagios
has started, parsed and loaded the configuration, triggered the "I'm
done now, modules" event, which causes merlin to start exporting the
timeperiod data from nagios memory, and then trigger ocimp to import it
into mysql.

Knowing when this is done is annoying. It was usually (but not always,
and that weird, random crash has been getting on my nerves for some
time) enough to just run the test on our RHEL machines. On our SLES
build slave, it wasn't, but with a one second sleep, everything works.

"I gave the drywall a nice, big slap, and the whole house didn't come
crashing down, so in other words, I've verified that everything is safe
and secure for habitation by humans"

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>