report_options: Use NOT, instead of empty(), when checking for emptyness
Because the array I'm checking is checked from inside an ArrayAccess
object rather than a proper array (yes, completely regular, proper
array, but I don't access it through a local variable), empty() will
always return TRUE, regardless of what the array actually contains.

This is the place in a commit message where I'd rant about the
staggering incompetence of whoever wrote this language, but I'm going to
save my energy, because I'll need it to debug code like this in the
future.

This actually fixes bug #6821.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>