Use "default" instead of "all" as keyword for default columns in system
Today, "all" and "default" are treated equally. But the difference is more
semantic. In the future, we may want to have a global configuraiton of default
columns, referring to which columns users should get when writing default. It
might relate to "all".

This patch makes it more future-proof if and when we want to implement that
feature. At that time, "all" and "default" won't be the same thing.

Signed-off-by: Max Sikström <msikstrom@op5.com>

Conflicts:

	src/generators/lsfilter/lscolumns.txt