ListView: time(column name) formats a unixtimestamp to a string
Custom columns is quite useful, specially for, well getting information like
last date checked. But it isn't that useful to get the value as a unix
timestamp. Not many people know unixtimestamps that well that they directly can
say which timestamp was yesterdey at lunch.

For those people that knows unixtimestamp doesn't need this function, but for
the rest of us, this is quite useful.

Signed-off-by: Max Sikström <msikstrom@op5.com>