countless-pagination: Explicitly forbid the next-link when not needed
A spider would follow the 'next'-links to the world's end. That's not
very useful.

The implementation is: check preconditions *outside* of the pagination
class and explicitly tell it when to skip the 'next' link.

The alternative would have been to declare a $count variable or such in
the pagination, and initialize it to 0 or so. That means I would need to
add a COUNT(*) result for the pagination to be able to calculate whether
the 'next' link is appropriate to print. That job is not worth it.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>