ListView: Don't use display_name... that will confuse users...
After a discussion with PS, mfalkvidd, and support, using display_name at all
is a bad idea. That hasn't been used before, and... well... don't show it.

We still change the behaviour that we remove the display_name column that
existed before, but it is possible to show it using custom columns, for users
who really need it.

Signed-off-by: Max Sikström <msikstrom@op5.com>