log: Make sure .logs have the right owner and group.
There may be some cases where a process that is run as root is the first to
write to a log file. In that case root is set as the owner of the log file
making everything else unable to log to it. With this change log sets proper
owners for the log file. Group is always set to monitor's primary group which is
the apache group making this change "sles-safe".

(This commit is imported from op5lib)

Signed-off-by: Henrik Smederod <hsmederod@op5.com>