authenticated: fallback to default skin if selected skins folder doesn't exists
we can come to a cornercase where a user might have a skin selected that doesn't
exists in the filesystem anymore, this makes a check and fallbacks to default/
skin if thats the case.

This leaves the user a little more userfriendly gui than a non css styled
op5 Monitor.

This closes bug #7242

Signed-off-by: Mattias Ryrlén <mattiasr@op5.com>