gui: Always display popups on hover, hardcode delay
Keeping the delay configurable (albeit hidden). This removes some of the
clutter in the user's own config.

Actually, the real reason for this commit is that the default value was
'off', allowing everybody to think the functionality was missing from
the beginning.

This has nothing to do with #3865 but its mentioned in its comments.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>