search: Make it possible to specify both s: and si:
Specifying queries with "si:" would remove anything that was defined
for hosts or services already. Instead, merge the two.

In a default monitor installation that isn't running on a Dell machine,
the query
 si: ERROR
should return all the failing Dell checks, and
 s: Merlin
should match all the merlin checks. However,
 s:Merlin and si:ERROR
should not match any of the Dell checks - only broken peers or pollers -
which this patch makes sure is what happens.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>