Nagioscmd: Use the correct interface both for extinfo and op5auth
It wasn't possible to check if a user was allowed for a commands for a service,
due to a variable type missmatch between extinfo controller and nagioscmd helper

Then a missmatch of names (host vs. hosts, service vs. services) between
authorized_for_object in op5auth and nagioscmd, efficiantly reduced the
permission so noone of the limited users had access to commands.

This resolves both. With a workaround, if using a new ninja on an old op5lib

Signed-off-by: Max Sikström <msikstrom@op5.com>