report_options: Set start & end time by inspecting input once
Report_options doesn't need to have the knowledge of cal_{start,end} or
time_{start,end} since they are decorators related to views and not
backend. By moving the logic to discover_options(), the *instance* of
Report_options can be happily unaware.

We're also decorating the start and end timestamps the same way we do
for summary report's options. And some cleanup of date format copy
pasting.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>