cucumber: Create HTML formated report of cuke output
Copied from nacoma, where the output is much more readable. Also for
consistency use the @skip tag for tests that should be "skipped", as
that name makes more sense - I like to use @todo for tests that aren't
good enough yet, as that name makes sense there. This is - again - how
nacoma also uses it.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>