Split report model into two models (three! three models!)
Before this, get_events was in the wrong place in Reports_Model. It
needed to be moved 4-5 methods down.

The fact that I can say that very seriously shows that something is
wrong with that model - it contains status methods, it contains summary
methods, and they have absolutely nothing whatsoever in common. Madness!

Put the status report code in a status report model, put the summary
report code in a summary report module, and the random helper methods
and variables that were hidden in the vast depths of infinitely long
methods gets to stay behind, hoping to provoke someone to do something
clever with them.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>