status_reports: Refactor state calculation
This lets me delete a dozen or so "do we have subreports, or are we one
ourselves?" conditionals that aren't trivial to wrap your head around.

It also extracts the state calculation itself - which is configurable an
varies between reports - from the task of fetching rows of status
data - which is always the same.

It also implements a calculator for the average value, meaning the
average now comes from the backend rather being something selected in
the view based on something calculated in the controller. We seem to be
fixing Yet Another component where we forgot about this every six months
or so (yarly - bug #6485 was literally fixed 6 months ago) - by putting
this in the backend, not only is bug #7315 fixed, but we also shouldn't
ever have to see this again.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>