reports: Merge "SLA Calculation Mode" with "Cluster mode"
These two options interacted in non-obvious ways, and having to explain
"It will use the average if that is selected from the SLA Calculation
mode, and you'll be able to tell by looking at the summary, where the
header 'SLA and average' will lie and only show an average. If not, it
will use the worst case or best case, depending on your cluster mode
setting, and you won't be able to tell at all, other than by looking at
the numbers in the report and manually calculating which one is
plausible - the reports will otherwise identical. Good luck!" grows old
quickly - people usually look confused by the time I get to "If not",
and by the time I've reach the end the confusion has usually sublimated
into sadness and frustration.

This fixes bug #7043.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>