Build a higher hierarchy of state calculators
Now each report contains a main, multi-object calculator, that
calculator contains a list of multi-object calculators (either one
multi-object calculator per group, or just one calculator), and this
middle calculator, in turn, contains the individual, per-object
calculators.

This means the feature request in bug #1751 can now be closed. It also
means I get to unify the code and remove a bunch of stuff no longer used
- in particular, I remove the logic to rename all the perfectly good
data we get from the model to other names, as none of the developers
who've touched both front and back end code in the reports - myself
included - seem to be able to make a mental model of where to put data
in the backend for the frontend to see it, or where to look in the
frontend to find backend data.

Sadness over our inability to keep form rendering consistent for
multi-object views caused me to generalize table rendering as well. Not
saying it's not still messy, just that there's fewer places to de-mess
it if someone wants to do that.

(I had plans to include a rewritten song lyric here, and even had some
draft lyrics jotted down, but after discovering @feministtswift, I've
decided not to even bother anymore - there's no reason to even keep
trying with something that brilliant out there)

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>