trends: Avoid rounding errors when there are many tiny events
Because the events were sized according to their duration, you could end
up in a situation where you had hundreds of events, all with a
ridiculously small width of 0.002% or something, wich was ignored both
by browsers and our PDF engine.

Fix by setting the start and stop percentages, and letting the rendering
engine auto-calculate the width as good it can. This means the graph
always starts at 0% and ends at 100%, and has a full band of colors
inbetween. Some events will still be rounded away, because we're subject
to the real world where pixels are discrete, but if a service is
flapping enough for this to be an issue, you'll likely still get a
green-red zebra stripy pattern.

This fixes bug #6656.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>