ListView: If link to filter with syntax error, display a default visual editor
When loading a link to a filter which has syntax errors, make sure the visual
editor actually shows a valid filter. Without that, it's impossible to actually
show any data without typing in the filter field.

This loades [hosts] all to the visual editor before loading the actual initial
filter.

Resolves #7363

Signed-off-by: Max Sikström <msikstrom@op5.com>