reports: Only exclude excluded time in piecharts
Timeperiods means the length of the period and the duration thereof are
not the same, so we cannot simply set unaccounted time as excluded, or
the percentages for the piecharts will be different from the report
percentages, which is exactly what the exclude-hack sought to prevent.

Instead, count excluded time as excluded, and other time not at all.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>