reports: Hide trends that are 100% OK
Trends are useful for correlating events and getting a basic overview,
however, if you have 100 services (which really isn't a lot) that are OK
all the time, and 5 services that you want to correlate with each other,
then the noise was way too high for the trends to help you.

By hiding the ones that work the way they should, it becomes easy to
concentrate on the services that give you trouble. As a bonus, it means
(if ever so slightly) less data to send to the client and render, making
trends a tiny bit faster as well.

This fixes bug #7473.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>