Auth: Dont allow login where no groups are defined
The authorization module did return if any groups were used in authorization,
but noone cared. It's actually just the permission flags that sets the
permission

But in some strange cases, where an LDAP user has the same name as a contact,
but shouldn't be that contact, it could get permissions to show status of those
objects that contact was contact for.

This shouldn't be a real issue, because noone should have name collissions
between contacts and user that isn't that contact. But anyway, just fail the
login if no groups are defined. It's better for everyone

Resolves #7490

(This commit is imported from op5lib)

Signed-off-by: Max Sikström <msikstrom@op5.com>