nagioscmd::is_authorized_for(): Throw exceptions if wanted
Introducing a new, third parameter should enscure that the old way of
doing things (letting the calling code have internal knowledge of this
method by knowing what error code means something) still works.

This is implemented so that the command part of the HTTP API (and, in
extension, everybody else) has an easier time submitting commands (which
is what this method really implies).

Signed-off-by: Carl Helmertz <chelmertz@op5.com>

Conflicts:
	application/helpers/nagioscmd.php